+++ lib/excoveralls/post.ex
@@ -16,7 +16,10 @@ defmodule ExCoveralls.Post do
     Jason.encode!(%{
       repo_token: options[:token],
       service_name: options[:service_name],
+      service_number: options[:service_number],
       source_files: source_info,
+      parallel: options[:parallel],
+      flag_name: options[:flagname],
       git: generate_git_info(options)
     })
   end