+++ lib/excoveralls/stats.ex
@@ -15,7 +15,7 @@ defmodule ExCoveralls.Stats do
 
   defmodule Line do
     @moduledoc """
-    Stores count information and source for a sigle line.
+    Stores count information and source for a single line.
     """
 
     defstruct coverage: nil, source: ""
@@ -94,6 +94,25 @@ defmodule ExCoveralls.Stats do
     Enum.map(stats, fn %{name: name} = stat ->
       %{stat | name: "#{apps_path}/#{sub_app_name}/#{name}"}
     end)
+  end
+
+  @doc """
+  Updates the paths to take into account the subdir and rootdir options
+  """
+  def update_paths(stats, options) do
+    sub_dir_set? = not (options[:subdir] in [nil, ""])
+    root_dir_set? = not (options[:rootdir] in [nil, ""])
+
+    cond do
+      sub_dir_set? ->
+        stats
+        |> Enum.map(fn m -> %{m | name: options[:subdir] <> Map.get(m, :name)} end)
+
+      root_dir_set? ->
+        stats
+        |> Enum.map(fn m -> %{m | name: String.trim_leading(Map.get(m, :name), options[:rootdir])} end)
+      true -> stats
+    end
   end
 
   @doc """