+++ test/post_test.exs
@@ -26,19 +26,31 @@ defmodule ExCoveralls.PostTest do
       Post.generate_json(@source_info, [
         token: "1234567890",
         service_name: "local",
+        service_number: "build_num_1",
         branch: "",
         committer: "",
         message: "",
-        sha: ""
+        sha: "",
+        flagname: "arbitrary_value"
       ])
 
-    assert json ==
-       "{\"git\":{\"branch\":\"\",\"head\":{\"committer_name\":\"\",\"id\":\"\",\"message\":\"\"}}," <>
-         "\"repo_token\":\"1234567890\"," <>
-         "\"service_name\":\"local\"," <>
-         "\"source_files\":" <>
-           "[{\"coverage\":[0,1,null,null]," <>
-             "\"name\":\"test/fixtures/test.ex\"," <>
-             "\"source\":\"defmodule Test do\\n  def test do\\n  end\\nend\\n\"}]}"
+    assert Jason.decode!(json) == %{
+      "flag_name" => "arbitrary_value",
+      "git" => %{
+        "branch" => "",
+        "head" => %{"committer_name" => "", "id" => "", "message" => ""}
+      },
+      "parallel" => nil,
+      "repo_token" => "1234567890",
+      "service_name" => "local",
+      "service_number" => "build_num_1",
+      "source_files" => [
+        %{
+          "coverage" => [0, 1, nil, nil],
+          "name" => "test/fixtures/test.ex",
+          "source" => "defmodule Test do\n  def test do\n  end\nend\n"
+        }
+      ]
+    }
   end
 end